refactor(cir): extract privilege service, clean schemas and format ui macs
