+++ .engram/manifest.json
@@ -72,6 +72,14 @@
       "sessions": 45,
       "memories": 154,
       "prompts": 103
+    },
+    {
+      "id": "ab90dbd5",
+      "created_by": "Alex",
+      "created_at": "2026-04-23T00:06:06Z",
+      "sessions": 8,
+      "memories": 42,
+      "prompts": 18
     }
   ]
 }
