+++ backend/app/routers/privilege_map.py
@@ -24,140 +24,14 @@ from app.core.security import get_current_active_user
 from app.core.rbac import require_roles, Role
 from app.core.limiter import limiter
 from app.services.audit import log_audit, EventCode
+from app.services.privilege_map_service import (
+    to_out_schema as _to_out,
+    compute_days_until_review,
+    raise_nas_conflict,
+    validate_segment_exception,
+)
 
 router = APIRouter(prefix="/privilege-map", tags=["privilege-map"])
-
-
-def raise_nas_conflict(payload: UserNasPrivilegeMapCreate):
-    if payload.nas_ip and payload.calling_station_id:
-        detail = f"Privilege map entry for user {payload.username} at NAS {payload.nas_ip} with MAC {payload.calling_station_id} already exists"
-    elif payload.calling_station_id:
-        detail = f"Privilege map entry for user {payload.username} with MAC {payload.calling_station_id} already exists"
-    elif payload.nas_ip:
-        detail = f"Privilege map entry for user {payload.username} at NAS {payload.nas_ip} already exists"
-    elif (
-        payload.segment_id is not None
-        and payload.target_start_ip is None
-        and payload.target_end_ip is None
-    ):
-        detail = "A base policy for this user and network segment already exists"
-    else:
-        detail = "Privilege map entry already exists"
-    
-    raise HTTPException(status_code=409, detail=detail)
-
-
-def _compute_days_until_review(review_date: Optional[datetime]) -> Optional[int]:
-    """Compute days until review. Negative means overdue."""
-    if review_date is None:
-        return None
-    today = datetime.utcnow().date()
-    delta = review_date.date() - today
-    return delta.days
-
-
-def _to_out(record: UserNasPrivilegeMap) -> UserNasPrivilegeMapOut:
-    """Serialize UserNasPrivilegeMap ORM → Out schema, resolving category name."""
-    days = _compute_days_until_review(record.review_date)
-    # Resolve category name if relationship is loaded
-    category_name = None
-    if record.nas_category_id is not None:
-        try:
-            category_name = record.category.name if record.category else None
-        except Exception:
-            category_name = None
-    return UserNasPrivilegeMapOut(
-        id=record.id,
-        username=record.username,
-        nas_ip=record.nas_ip,
-        calling_station_id=record.calling_station_id,
-        nas_category_id=record.nas_category_id,
-        nas_category_name=category_name,
-        nas_identifier=record.nas_identifier,
-        nas_vendor=record.nas_vendor,
-        radius_group=record.radius_group,
-        privilege_level=record.privilege_level,
-        justification=record.justification,
-        approved_by=record.approved_by,
-        review_date=record.review_date.date() if record.review_date else None,
-        is_active=record.is_active,
-        created_at=record.created_at,
-        updated_at=record.updated_at,
-        days_until_review=days,
-        segment_id=record.segment_id,
-        segment_name=record.segment.name if record.segment else None,
-        target_start_ip=record.target_start_ip,
-        target_end_ip=record.target_end_ip,
-    )
-
-
-async def validate_segment_exception(
-    db: AsyncSession,
-    payload: UserNasPrivilegeMapCreate,
-    exclude_id: Optional[int] = None,
-):
-    if payload.segment_id is None:
-        return
-
-    segment_res = await db.execute(
-        select(NetworkSegment).where(NetworkSegment.id == payload.segment_id)
-    )
-    segment = segment_res.scalars().first()
-    if not segment:
-        raise HTTPException(status_code=404, detail="Network segment not found")
-
-    if payload.target_start_ip and payload.target_end_ip:
-        start_ip = ipaddress.ip_address(payload.target_start_ip)
-        end_ip = ipaddress.ip_address(payload.target_end_ip)
-        segment_net = ipaddress.ip_network(segment.cidr, strict=False)
-
-        if start_ip not in segment_net or end_ip not in segment_net:
-            raise HTTPException(
-                status_code=422,
-                detail="Exception IPs must strictly fall within the parent NetworkSegment CIDR",
-            )
-
-        stmt = select(UserNasPrivilegeMap).where(
-            and_(
-                UserNasPrivilegeMap.username == payload.username,
-                UserNasPrivilegeMap.segment_id == payload.segment_id,
-                UserNasPrivilegeMap.target_start_ip.is_not(None),
-            )
-        )
-        if exclude_id is not None:
-            stmt = stmt.where(UserNasPrivilegeMap.id != exclude_id)
-
-        existing_res = await db.execute(stmt)
-        existing = existing_res.scalars().all()
-
-        for exc in existing:
-            exc_start = ipaddress.ip_address(exc.target_start_ip)
-            exc_end = ipaddress.ip_address(exc.target_end_ip)
-            if start_ip <= exc_end and end_ip >= exc_start:
-                raise HTTPException(
-                    status_code=422,
-                    detail=f"IP range overlaps with existing exception: {exc.target_start_ip} - {exc.target_end_ip}",
-                )
-        return
-
-    stmt = select(UserNasPrivilegeMap).where(
-        and_(
-            UserNasPrivilegeMap.username == payload.username,
-            UserNasPrivilegeMap.segment_id == payload.segment_id,
-            UserNasPrivilegeMap.target_start_ip.is_(None),
-            UserNasPrivilegeMap.target_end_ip.is_(None),
-        )
-    )
-    if exclude_id is not None:
-        stmt = stmt.where(UserNasPrivilegeMap.id != exclude_id)
-
-    existing_base_res = await db.execute(stmt)
-    existing_base = existing_base_res.scalars().first()
-    if existing_base:
-        raise HTTPException(
-            status_code=409,
-            detail="A base policy for this user and network segment already exists",
-        )
 
 
 @router.get("", response_model=list[UserNasPrivilegeMapOut])
@@ -241,7 +115,18 @@ async def create_privilege_map_bulk(
         db.add(record)
         records.append(record)
 
-    await db.commit()
+    try:
+        await db.commit()
+    except IntegrityError:
+        await db.rollback()
+        # Create a dummy create-payload for conflict messaging
+        dummy = UserNasPrivilegeMapCreate(
+            username=payload.username,
+            nas_ip=payload.nas_ips[0],
+            radius_group=payload.radius_group
+        )
+        raise_nas_conflict(dummy)
+
     for record in records:
         await db.refresh(record)
 
@@ -273,10 +158,11 @@ async def create_privilege_map_category(
         payload.nas_category_id is None
         and payload.segment_id is None
         and payload.nas_ip is None
+        and payload.calling_station_id is None
     ):
         raise HTTPException(
             status_code=422,
-            detail="A targeting method (nas_ip, nas_category_id, or segment_id) is required",
+            detail="A targeting method (nas_ip, calling_station_id, nas_category_id, or segment_id) is required",
         )
 
     await validate_segment_exception(db, payload)
